Move how we handle `searchParams` for WidgetHashHandler
